refactor: ConfigMapSecret folder renamed -> ConfigMapSecretOld
