+++ src/components/v2/values/chartValuesDiff/ChartValuesView.tsx
@@ -137,7 +137,7 @@ import {
     EMPTY_YAML_ERROR,
 } from './ChartValuesView.constants'
 import ClusterNotReachableDailog from '../../../common/ClusterNotReachableDailog/ClusterNotReachableDialog'
-import { VIEW_MODE } from '@Pages/Shared/ConfigMapSecret/Secret.utils'
+import { VIEW_MODE } from '@Pages/Shared/ConfigMapSecretOld/Secret.utils'
 import IndexStore from '../../appDetails/index.store'
 import { AppDetails } from '../../appDetails/appDetails.type'
 import { AUTO_GENERATE_GITOPS_REPO, CHART_VALUE_ID } from './constant'
